De-sloppification: deduplicate BlockingLoadErrorState and getBlockingLoadErrorMessage helpers (finding #2 only). Centralised in shared error helper; updated all consumers. No change to agent model/config drift (finding #1 left untouched).
